replace parseInt with parseFloat

+++ date_calculator.js
@@ -1,27 +1,27 @@
-let c=parseInt(prompt("Enter the 'FIRST TWO DIGITS' of the year you were born"));
+let c=parseFloat(prompt("Enter the 'FIRST TWO DIGITS' of the year you were born"));
 let cRange= Array.from({length:99}, (_, index) => index + 1);
 while(cRange.includes(c)===false)
     {
         alert("Please enter a valid number. MUST be an integer, at least 16 and at most 99");
-        c=parseInt(prompt("Enter the 'FIRST TWO DIGITS' of the year you were born"));
+        c=parseFloat(prompt("Enter the 'FIRST TWO DIGITS' of the year you were born"));
     }
 
 
-let y=parseInt(prompt("Enter the 'LAST TWO DIGITS' of the year youy were born"));
+let y=parseFloat(prompt("Enter the 'LAST TWO DIGITS' of the year youy were born"));
 let yRange=Array.from({length:99}, (_, index) => index + 1);
 while(cRange.includes(y)===false)
     {
         alert("Please enter a valid year. MUST be an integer, at least 0 and at most 99");
-        y=parseInt(prompt("Enter the 'LAST TWO DIGITS' of the year youy were born"));
+        y=parseFloat(prompt("Enter the 'LAST TWO DIGITS' of the year youy were born"));
     }
 
 
-let m=parseInt(prompt("Enter the 'MONTH' number you were born"));
+let m=parseFloat(prompt("Enter the 'MONTH' number you were born"));
 let mRange=Array.from({length:12}, (_, index) => index + 1);
 while(cRange.includes(y)===false)
     {
         alert("Please enter a valid month. MUST be an integer, at least 1 and at most 12");
-        m=parseInt(prompt("Enter the 'MONTH' number you were born"));
+        m=parseFloat(prompt("Enter the 'MONTH' number you were born"));
     }
 
 
@@ -29,7 +29,7 @@ let N=31
 let dayRange= Array.from({length:N}, (_, index) => index + 1);
 let longMonths=[1,3,5,7,8,10,12];
 let shortMonths=[4,6,9,11];
-let d=parseInt(prompt("Enter the 'DATE' of the month you were born: Can be from 1 through 31"));
+let d=parseFloat(prompt("Enter the 'DATE' of the month you were born: Can be from 1 through 31"));
 if (longMonths.includes(m))
     {
         N=31
@@ -37,7 +37,7 @@ if (longMonths.includes(m))
         while(dayRange.includes(d)===false)
         {
             alert("Please enter a valid day. MUST be an integer, at least 1 and at most 31")
-            d=parseInt(prompt("Enter the 'DAY' of the month you were born"));
+            d=parseFloat(prompt("Enter the 'DAY' of the month you were born"));
         }
     }
 else if(shortMonths.includes(m))
@@ -47,7 +47,7 @@ else if(shortMonths.includes(m))
         while(dayRange.includes(d)===false)
         {
             alert("Please enter a valid day. MUST be an integer, at least 1 and at most 30")
-            d=parseInt(prompt("Enter the 'DAY' of the month you were born"));
+            d=parseFloat(prompt("Enter the 'DAY' of the month you were born"));
         }
     }
 else
@@ -61,7 +61,7 @@ else
                         while(dayRange.includes(d)===false)
                         {
                             alert("Please enter a valid day. MUST be an integer, at least 1 and at most 29, since it is February")
-                            d=parseInt(prompt("Enter the 'DAY' of the month you were born"));
+                            d=parseFloat(prompt("Enter the 'DAY' of the month you were born"));
                         }
                     }
                 else
@@ -71,7 +71,7 @@ else
                         while(dayRange.includes(d)===false)
                         {
                             alert("Please enter a valid day. MUST be an integer, at least 1 and at most 28, since it is February in an odinary year")
-                            d=parseInt(prompt("Enter the 'DAY' of the month you were born"));
+                            d=parseFloat(prompt("Enter the 'DAY' of the month you were born"));
                         }
                     }
                 
@@ -85,7 +85,7 @@ else
                     while(dayRange.includes(d)===false)
                         {
                             alert("Please enter a valid day. MUST be an integer, at least 1 and at most 29, since it is February")
-                            d=parseInt(prompt("Enter the 'DAY' of the month you were born"));
+                            d=parseFloat(prompt("Enter the 'DAY' of the month you were born"));
                         }
                 }
             else
@@ -95,7 +95,7 @@ else
                     while(dayRange.includes(d)===false)
                         {
                             alert("Please enter a valid day. MUST be an integer, at least 1 and at most 28, since it is February in an odinary year")
-                            d=parseInt(prompt("Enter the 'DAY' of the month you were born"));
+                            d=parseFloat(prompt("Enter the 'DAY' of the month you were born"));
                         }
                 }
         }
